+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [unreleased]
+
+### Added
+
+- Added support for subframe interpolation of per-frame collision boundaries.
+
+
 ## [0.13.0-exp.1] - 2024-04-16
 
 ### Added